refactor: toString test helper
