+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMultiSetImpl.kt
@@ -1,0 +1,9 @@
+package xyz.lbres.kotlinutils.set.multiset.const
+
+import xyz.lbres.kotlinutils.set.multiset.MultiSet
+
+/**
+ * [MultiSet] implementation where values of elements are assumed to be constant.
+ * Behavior is not defined if values of elements are changed (i.e. elements are added to a mutable list).
+ */
+internal class ConstMultiSetImpl<E>(initialElements: Collection<E>) : ConstMultiSet<E>(initialElements)
