+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MultiSetImpl.kt
@@ -9,10 +9,10 @@ internal class MultiSetImpl<E> : AbstractMultiSetImpl<E> {
     /**
      * Initialize set from a collection of values.
      */
-    internal constructor(elements: Collection<E>) : super(elements)
+    constructor(elements: Collection<E>) : super(elements)
 
     /**
      * Initialize set from existing counts.
      */
-    internal constructor(counts: Map<E, Int>) : super(counts)
+    constructor(counts: Map<E, Int>) : super(counts)
 }
