+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MutableMultiSetImplTest.kt
@@ -88,39 +88,7 @@ class MutableMultiSetImplTest {
 
     @Test
     fun testToString() {
-        var set: MutableMultiSet<Int> = mutableMultiSetOf()
-        var expected = "[]"
-        assertEquals(expected, set.toString())
-
-        set = mutableMultiSetOf(100, -200, 400, 800)
-        expected = "[100, -200, 400, 800]"
-        assertEquals(expected, set.toString())
-
-        set = mutableMultiSetOf(1, 1, 1, 1)
-        expected = "[1, 1, 1, 1]"
-        assertEquals(expected, set.toString())
-
-        set = mutableMultiSetOf(2, 4, 2, 1)
-        expected = "[2, 4, 2, 1]"
-        assertEquals(expected, set.toString())
-
-        set.add(5)
-        expected = "[2, 4, 2, 1, 5]"
-        assertEquals(expected, set.toString())
-
-        set.remove(2)
-        expected = "[4, 2, 1, 5]"
-        assertEquals(expected, set.toString())
-
-        val mutableList1 = mutableListOf(1, 2, 3)
-        val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
-
-        expected = "[[1, 2, 3], [0, 5, 7]]"
-        assertEquals(expected, listSet.toString())
-
-        mutableList2.clear()
-        expected = "[[1, 2, 3], []]"
-        assertEquals(expected, listSet.toString())
+        runMultiSetMutableToStringTests(createSet(), createSet())
+        runMultiSetMutableElementToStringTests(createSet())
     }
 }
